Make test for truthy more robust
Conditional tests would evaluate to 'true' when a user passed the
variable in the extra-vars. Regardless of what was actually specified by
the user. Because the value would be seen as a string, and a variable
having content evaluates to true unless you add the 'bool' filter.

This commit adds the 'bool' filter to the conditional tests. This
improves the role to behave more in line with expectations when using
extra-vars.